[Refactor] 불필요한 주석 삭제 및 폴더 구조 조정
